+++ src/Entities/Schema.php
@@ -135,6 +135,9 @@ final class Schema implements IteratorAggregate, ArrayAccess
             'subscriptions/{subscriptionId}/upcoming-invoices' => function (array $content) {
                 return new Collection(new Invoice(), $content);
             },
+            'subscriptions/{subscriptionId}/upcoming-invoice' => function (array $content) {
+                return new Invoice($content);
+            },
             'subscriptions/{subscriptionId}/lead-source' => function (array $content) {
                 return new LeadSource($content);
             },