+++ src/Services/SubscriptionService.php
@@ -170,11 +170,26 @@ final class SubscriptionService extends Service
     }
 
     /**
+     * @deprecated Use self::getUpcomingInvoice()
+     *
      * @param string $subscriptionId
      *
      * @return Invoice[]|Collection
      */
     public function getUpcomingInvoices($subscriptionId)
+    {
+        return $this->client()->get(
+            'subscriptions/{subscriptionId}/upcoming-invoices',
+            ['subscriptionId' => $subscriptionId]
+        );
+    }
+
+    /**
+     * @param string $subscriptionId
+     *
+     * @return Invoice
+     */
+    public function getUpcomingInvoice($subscriptionId)
     {
         return $this->client()->get(
             'subscriptions/{subscriptionId}/upcoming-invoices',