+++ src/Entities/PaymentInstruments/AchInstrument.php
@@ -21,11 +21,32 @@ use Rebilly\Entities\PaymentMethodInstrument;
 class AchInstrument extends PaymentMethodInstrument
 {
     /**
+     * {@inheritdoc}
+     */
+    public function __construct(array $data = [])
+    {
+        if (isset($data['bankAccountId'])) {
+            $data['paymentInstrumentId'] = $data['bankAccountId'];
+            unset($data['bankAccountId']);
+        }
+        parent::__construct(['method' => $this->methodName()] + $data);
+    }
+
+    /**
+     * @return string
+     */
+    public function getPaymentInstrumentId()
+    {
+        return $this->getAttribute('paymentInstrumentId');
+    }
+
+    /**
+     * @deprecated
      * @return string
      */
     public function getBankAccountId()
     {
-        return $this->getAttribute('bankAccountId');
+        return $this->getPaymentInstrumentId();
     }
 
     /**
@@ -33,9 +54,20 @@ class AchInstrument extends PaymentMethodInstrument
      *
      * @return $this
      */
+    public function setPaymentInstrumentId($value)
+    {
+        return $this->setAttribute('paymentInstrumentId', $value);
+    }
+
+    /**
+     * @deprecated
+     * @param string $value
+     *
+     * @return $this
+     */
     public function setBankAccountId($value)
     {
-        return $this->setAttribute('bankAccountId', $value);
+        return $this->setPaymentInstrumentId($value);
     }
 
     /**