+++ CHANGELOG.md
@@ -27,8 +27,7 @@ Security - in case of vulnerabilities.
 - [x] Added `AchInstrument::setPaymentInstrumentId` method
 - [x] Added `PayPalInstrument::getPaymentInstrumentId` method
 - [x] Added `PayPalInstrument::setPaymentInstrumentId` method
-- [x] Added **Cashier Requests API**
-- [x] Added **Cashier Strategy API**
+- [x] Added **Cashier API**
 
 ### Deprecated